realtime_voice/requirements.txt
# Real-Time Voice Assistant Dependencies

# Core
requests>=2.28.0
psycopg2-binary>=2.9.0

# Speech-to-Text
faster-whisper>=0.9.0

# Voice Activity Detection (optional but recommended)
# silero-vad  # Loaded via torch.hub

# PyTorch (for VAD and some models)
torch>=2.0.0

# Audio processing
numpy>=1.24.0
soundfile>=0.12.0

# Gradio (for web UI)
gradio>=4.0.0

# TTS is expected via Databricks Model Serving endpoint.
